+++ src/wininput.c
@@ -1315,6 +1315,12 @@ toggle_bidi()
   term.disable_bidi = !term.disable_bidi;
 }
 
+void
+toggle_dim_margins()
+{
+  term.dim_margins = !term.dim_margins;
+}
+
 static void scroll_HOME()
   { SendMessage(wnd, WM_VSCROLL, SB_TOP, 0); }
 static void scroll_END()
@@ -1622,6 +1628,12 @@ mflags_bidi()
 }
 
 static uint
+mflags_dim_margins()
+{
+  return term.dim_margins ? MF_CHECKED : MF_UNCHECKED;
+}
+
+static uint
 mflags_options()
 {
   return config_wnd ? MF_GRAYED : MF_ENABLED;
@@ -1736,6 +1748,7 @@ static struct function_def cmd_defs[] = {
   {"toggle-auto-repeat", {.fct = toggle_auto_repeat}, mflags_auto_repeat},
   {"toggle-bidi", {.fct = toggle_bidi}, mflags_bidi},
   {"refresh", {.fct = refresh}, 0},
+  {"toggle-dim-margins", {.fct = toggle_dim_margins}, mflags_dim_margins},
 
   {"super", {.fct_key = super_down}, 0},
   {"hyper", {.fct_key = hyper_down}, 0},